+++ Changelog.md
@@ -75,6 +75,7 @@
  + Updated first floor boss moves
 # - Alpha v4.2
  + Added a splash screen on startup
+ + Added new 'talk' dialogue to floor 2 arena.
 # Known issues:
  - Update current song after victory
 # TODO: